feat: add a release workflow (#38)
This is a first attempt at a release workflow. I can't test it since it
won't be available on the default branch until after merging this.
I don't know what secrets this repo will have access to. We will need a
cargo access token, which should probably be an org wide secret.

The workflow will be triggered manually. When run it will result in the
following:
- creates a git tag with the version
- publishes to crates.io via `cargo publish`
- creates a release on github

This is part of https://github.com/recallnet/rust-recall/issues/17